src/requirements.txt
allure-pytest
pytest
selenium
webdriver_manager